+++ lib/fb_news_feed.dart
@@ -5,18 +5,16 @@ import 'dart:convert';
 import 'package:flutter/material.dart';
 
 // Project imports:
-import '../flutter_fb_news.dart';
 import 'fb_news_attachments_photos.dart';
 import 'fb_news_attachments_videos.dart';
 import 'fb_news_footer.dart';
 import 'fb_news_header.dart';
 import 'fb_news_message.dart';
+import 'flutter_fb_news.dart';
+import 'flutter_fb_news_config.dart';
 
 /// Only for internal use of flutter_fb_news
 class FbNewsFeed extends StatelessWidget {
-  /// Only for internal use of flutter_fb_news
-  final String subtitle;
-
   /// Only for internal use of flutter_fb_news
   final String feedResponse;
 
@@ -24,29 +22,16 @@ class FbNewsFeed extends StatelessWidget {
   final String profilePictureUrl;
 
   /// Only for internal use of flutter_fb_news
-  final List<FbNewsFieldName> fields;
-
-  /// Only for internal use of flutter_fb_news
-  final Color borderColor;
-
-  /// Only for internal use of flutter_fb_news
-  final Color backgroundColor;
-
-  /// Only for internal use of flutter_fb_news
-  final Color textColor;
+  final FbNewsConfig config;
 
   const FbNewsFeed({
     @required this.feedResponse,
-    @required this.subtitle,
     @required this.profilePictureUrl,
-    @required this.fields,
-    this.borderColor,
-    this.backgroundColor,
-    this.textColor,
+    @required this.config,
   });
 
   bool hasField(String internalKey) {
-    return fields
+    return config.fields
         .where((element) => element.internalKey == internalKey)
         .isNotEmpty;
   }
@@ -60,12 +45,12 @@ class FbNewsFeed extends StatelessWidget {
       children: jsonDecode(feedResponse)["data"]
           .map(
             (feed) => Card(
-              color: borderColor ?? Theme.of(context).accentColor,
+              color: config.borderColor ?? Theme.of(context).accentColor,
               child: Container(
                 width: 460,
                 padding: EdgeInsets.all(10),
                 child: Card(
-                  color: backgroundColor,
+                  color: config.backgroundColor,
                   child: Container(
                     width: 430,
                     child: Column(
@@ -74,10 +59,9 @@ class FbNewsFeed extends StatelessWidget {
                           FbNewsFields.header.internalKey,
                         )
                             ? FbNewsHeader(
-                                subtitle: subtitle,
                                 feed: feed,
                                 profilePictureUrl: profilePictureUrl,
-                                textColor: textColor,
+                                config: config,
                               )
                             : Container(),
                         hasField(
@@ -85,7 +69,7 @@ class FbNewsFeed extends StatelessWidget {
                         )
                             ? FbNewsMessage(
                                 feed: feed,
-                                textColor: textColor,
+                                config: config,
                               )
                             : Container(),
                         hasField(
@@ -113,7 +97,7 @@ class FbNewsFeed extends StatelessWidget {
                         )
                             ? FbNewsFooter(
                                 feed: feed,
-                                textColor: textColor,
+                                config: config,
                               )
                             : Container(),
                       ],