+++ lib/fb_news_footer.dart
@@ -4,16 +4,20 @@ import 'package:flutter/material.dart';
 // Package imports:
 import 'package:url_launcher/url_launcher.dart';
 
+// Project imports:
+import 'flutter_fb_news_config.dart';
+
 /// Only for internal use of flutter_fb_news
 class FbNewsFooter extends StatelessWidget {
   /// Only for internal use of flutter_fb_news
   final Map<String, dynamic> feed;
 
-  /// Only for internal use of flutter_fb_news
-  final Color textColor;
+  /// Only for internal use of flutter_fb_newsconfig
+  final FbNewsConfig config;
+
   const FbNewsFooter({
     @required this.feed,
-    this.textColor,
+    this.config,
   });
 
   @override
@@ -26,7 +30,7 @@ class FbNewsFooter extends StatelessWidget {
               children: [
                 Icon(
                   Icons.thumb_up,
-                  color: textColor,
+                  color: config.textColor,
                 ),
                 SizedBox(
                   width: 5,
@@ -34,7 +38,7 @@ class FbNewsFooter extends StatelessWidget {
                 Text(
                   feed["likes"]["data"].length.toString(),
                   style: TextStyle(
-                    color: textColor,
+                    color: config.textColor,
                   ),
                 ),
               ],
@@ -45,7 +49,7 @@ class FbNewsFooter extends StatelessWidget {
             .toString()
             .substring(0, 16),
         style: TextStyle(
-          color: textColor,
+          color: config.textColor,
         ),
       ),
       onTap: () {