+++ test/widgets/fb_news_footer_test.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 // Project imports:
-import 'package:flutter_fb_news/widgets/fb_news_footer.dart';
+import 'package:flutter_fb_news/fb_news_footer.dart';
 import '../TestMaterialWidget.dart';
 
 void main() {