+++ test/widgets/fb_news_message_test.dart
@@ -2,7 +2,7 @@
 import 'package:flutter_test/flutter_test.dart';
 
 // Project imports:
-import 'package:flutter_fb_news/widgets/fb_news_message.dart';
+import 'package:flutter_fb_news/fb_news_message.dart';
 import '../TestMaterialWidget.dart';
 
 void main() {